+++ UserManagementApi/Repositories/ICustomerRepository.cs
@@ -6,9 +6,9 @@ namespace UserManagementApi.Repositories
     public interface ICustomerRepository
     {
         Task<IEnumerable<Customer>> GetAll();
-        Task<Customer> GetCustomer(long id);
+        Task<Customer> GetCustomer(int id);
         Task<IEnumerable<Customer>> AddCustomer(Customer customer);
         Task<IEnumerable<Customer>> UpdateCustomer(Customer customer);
-        Task<Customer> DeleteCustomer(long id);
+        Task<Customer> DeleteCustomer(int id);
     }
 }
